web: daily notes, simplified — one rule per decision (hickey + lowy)
Four findings, all of them one thing said twice.

`markdown/follow.ts` is gone and `followed` lives beside `<Link>` in
router.tsx. The two were the same decision — a click becomes a route — with the
modifier/button guard spelled once in each, and that guard is now one `ours`
predicate both read: what a reader meant by a keypress is not a rule to keep in
two heads, and the only difference between the two shapes is who wrote the
markup. The module header already claimed the subject ("the one component
allowed to change the address"), which is the placement test answering itself.
Its note now also records why the listener sits on the PANE rather than on
`document`: Solid dispatches a descendant's handler before an ancestor's, so a
`<Link>` inside the pane has provably run first, where two document listeners
would be ordered by registration and losing that race is one click pushing two
history entries.

`noteDateOf` leaves the package's public index. Nothing outside `format` calls
it — the web reads `dailyNoteDays` and `dailyNotesOn` — and the index's own
header says the spellings a rule happens to use are not contract. One socket,
not three wires.

`relative` moves from the two marks into `BOX`. Day.tsx's own note says every
property the marks touch is decided in exactly one of them, and the corner fold
had made `position` the exception.

And `noteDateOf` stops spelling `.md` a second time: `fileKind` already owns
what a document is, so the extension is cut by finding it.
